Make admin signing approvals and rejections testable

+++ app/models/alumvest/signing/base.rb
@@ -5,17 +5,21 @@ module Alumvest::Signing
     include Alumvest::Signing::State
 
     def remove_investor_reviewer
-      # TODO: let envelope do this
-      client = DocusignRest::Client.new
-      envelope_id = envelope.envelope_id
-      recipients = client.get_envelope_recipients(:envelope_id => envelope_id)
-      recipient_id = recipients['signers'].select{|s| s["roleName"] == "investor-reviewer"}.first["recipientId"]
-      result = client.delete_envelope_recipient({:envelope_id => envelope_id, :recipient_id => recipient_id})
-      begin
-        result['signers'].first['status'] == 'deleted'
-      rescue
-        # TODO: log some stuff, eh?
-        false
+      unless Rails.env.test?
+        # TODO: this belongs in envelope#talks_to_docusign
+        client = DocusignRest::Client.new
+        envelope_id = envelope.envelope_id
+        recipients = client.get_envelope_recipients(:envelope_id => envelope_id)
+        recipient_id = recipients['signers'].select { |s| s["roleName"] == "investor-reviewer" }.first["recipientId"]
+        result = client.delete_envelope_recipient({:envelope_id => envelope_id, :recipient_id => recipient_id})
+        begin
+          result['signers'].first['status'] == 'deleted'
+        rescue
+          # TODO: log some stuff, eh?
+          false
+        end
+      else # we are in test
+        true
       end
     end
   end
